+++ Sources/NodeSimEx/Export_UserDll.h
@@ -74,11 +74,12 @@ typedef void (*SETKEYPRESSED)(KEYPRESSED);
 typedef int (*GETMESSAGENAME)(DWORD, DWORD , char* ,DWORD );
 typedef void (*DLLGETMESSAGENAME)(GETMESSAGENAME);
 
-typedef long (*TIMENOW)();
+typedef long long (*TIMENOW)();
 typedef void (*DLLTIMENOW)(TIMENOW);
 
 typedef DWORD (*GETFIRSTCANDBNAME)(char [], DWORD);
 typedef void (*DLLGETFIRSTCANDBNAME)(GETFIRSTCANDBNAME);
+
 // Name of the exported function from user-defined DLL
 #define NAME_FUNC_LOG_DISABLE           "vSetDisableLoggingProcAddress"
 #define NAME_FUNC_LOG_ENABLE            "vSetEnableLoggingProcAddress"
